fix(grass): cast narrowing in LOD counter copy
const D3D11_BOX countBox{ countOffset, 0, 0, countOffset + sizeof(uint32_t), 1, 1 }
narrows countOffset + sizeof(uint32_t) (promoted to size_t by sizeof's
return type) into the UINT-typed 4th field inside a brace-init list --
ill-formed C++, but cl.exe silently accepts it while clang-cl (the
Linux cross-compile toolchain) rejects it with -Wc++11-narrowing.
Wrapped in an explicit cast.

Found by a peer session (commonlibvr-c2) reviewing the PR's failing
Linux cross-compile job.

+++ src/Features/GrassOptimizations.cpp
@@ -848,7 +848,7 @@ void GrassOptimizations::CullBucket(GrassBucket& b, ID3D11DeviceContext* ctx)
 			continue;
 		for (uint32_t eye = 0; eye < eyeCount; ++eye) {
 			const UINT countOffset = LODCounterOffsetForEye(eye, tier);
-			const D3D11_BOX countBox{ countOffset, 0, 0, countOffset + sizeof(uint32_t), 1, 1 };
+			const D3D11_BOX countBox{ countOffset, 0, 0, static_cast<UINT>(countOffset + sizeof(uint32_t)), 1, 1 };
 			ctx->CopySubresourceRegion(bin.argsBuf, 0, InstanceCountOffsetForEye(eye), 0, 0, b.lodCounterBuf, 0, &countBox);
 		}
 	}
